pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "avalanche-gpt"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "Boje Deforce", email = "bojedeforce@gmail.com" },
  { name = "Martijn Sterkenburg", email = "sterkenburgmartijn@gmail.com" },  
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "beartype",
  "requests",
  "python-dotenv",
  "openai",
  "tiktoken" # For token counting, may come in useful
]

[project.urls]
Documentation = "https://github.com/strong-force-projects/avalanche-gpt#readme"
Issues = "https://github.com/strong-force-projects/avalanche-gpt/issues"
Source = "https://github.com/strong-force-projects/avalanche-gpt"

[tool.hatch.version]
path = "src/avalanche_gpt/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/avalanche_gpt tests}"

[tool.coverage.run]
source_pkgs = ["avalanche-gpt", "tests"]
branch = true
parallel = true
omit = [
  "src/avalanche_gpt/__about__.py",
]

[tool.coverage.paths]
avalanche-gpt = ["src/avalanche_gpt", "*/avalanche-gpt/src/avalanche_gpt"]
tests = ["tests", "*/avalanche-gpt/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
